Change timline in Store.transactions
