+++ src/components/tables/TableListView.vue
@@ -125,13 +125,13 @@ export default {
       type: Object
     },
 
-    timelineNextAction: {
-      type: String
-    },
+    // timelineNextAction: {
+    //   type: String
+    // },
 
-    timelinePreviousAction: {
-      type: String
-    },
+    // timelinePreviousAction: {
+    //   type: String
+    // },
 
     pageSize: {
       type: Number,
@@ -210,7 +210,8 @@ export default {
     nextPage() {
       if (this.nextPageExist) {
         if (this.timelinePagination)
-          this.$store.dispatch(this.timelineNextAction)
+          //this.$store.dispatch(this.timelineNextAction)
+          this.timeline.fetchNext()
         else
           this.pageIndex++
       }
@@ -219,7 +220,8 @@ export default {
     prevPage() {
       if (this.prevPageExist) {
         if (this.timelinePagination)
-          this.$store.dispatch(this.timelinePreviousAction)
+          //this.$store.dispatch(this.timelinePreviousAction)
+          this.timeline.fetchPrevious()
         else
           this.pageIndex--
       }
