+++ src/infrastructure/getTransaction.js
@@ -139,7 +139,15 @@ class sdkTransaction {
     const response = await axios.get(http.nodeUrl + path)
     const transactions = response.data.map(info => dto.createTransactionFromDTO(info, http.networkType))
 
-    return format.formatTransactions(transactions)
+    return format.formatTransactions(transactions).map(el => ({
+      height: el.blockHeight,
+      deadline: el.deadline,
+      transactionHash: el.transactionHash,
+      type: el.transactionBody?.type,
+      fee: el.fee,
+      signer: el.signer,
+      recipient: el.transactionBody?.recipient
+    }))
   }
 
   static getTransactionsSinceHashWithLimit = async (limit, transactionType, sinceHash) => {
