+++ src/store/manager/Filter.js
@@ -17,30 +17,33 @@
  */
 
 export default class Filter {
-    constructor(name, options, store) {
-        if(typeof name === 'string')
+    constructor(name, options) {
+        if(typeof name !== 'string')
             throw Error('Failed to construct Filter. Name is not provided');
-        if(typeof options === 'object')
+        if(typeof options !== 'object')
             throw Error('Failed to construct Filter. Options map is not provided');
-        if(!(store && store.state && store.getters && store.commit && store.dispatch))
-            throw Error('Failed to construct Filter. Store context is not provided');
-        
-        this.loading = false;
-        this.error = false;
+
         this.name = name;
         this.options = options;
-        this.store = store;
         this.current = Object.keys(options)[0]
+        this.store = {};
     }
+
+    setStore(store) {
+        this.store = store;
+        this.store.dispatch(this.name, this);
+        return this;
+    }
+
     get currentManager() {
-        return this.getters[this.current] || {};
+        return this.store?.getters ? this.store.getters[this.current] || {} : {};
     }
 
     get filterOptions() {
         return this.options;
     }
 
-    get currentOption() {
+    get filterValue() {
         return this.current
     }
 
@@ -49,11 +52,11 @@ export default class Filter {
     }
 
     get error() {
-        return typeof this.currentManager.error === 'boolean' ? this.currentManager.error : true;
+        return typeof this.currentManager.error === 'boolean' ? this.currentManager.error : false;
     }
 
     get loading() {
-        return typeof this.currentManager.loading === 'boolean' ? this.currentManager.loading : false;
+        return typeof this.currentManager.loading === 'boolean' ? this.currentManager.loading : true;
     }
 
     get canFetchPrevious() {
@@ -68,25 +71,32 @@ export default class Filter {
         return typeof this.currentManager.index === 'number' ? this.currentManager.index : 0;
     }
 
+    initialFetch() {
+        if(typeof this.currentManager.initialFetch === 'function')
+            return this.currentManager.initialFetch();
+        else 
+            console.error('Cannot call "initialFetch" method of', this.current)
+    }
+
     fetchNext() {
         if(typeof this.currentManager.fetchNext === 'function')
-            return this.fetchNext();
+            return this.currentManager.fetchNext();
         else 
             console.error('Cannot call "fetchNext" method of', this.current)
     }
 
     fetchPrevious() {
         if(typeof this.currentManager.fetchPrevious === 'function')
-            return this.fetchPrevious();
+            return this.currentManager.fetchPrevious();
         else 
             console.error('Cannot call "fetchPrevious" method of', this.current)
     }
 
-    selectOption(option) {
+    changeFilterValue(option) {
         this.current = option;
 
         if(typeof this.currentManager.initialFetch === 'function')
-            currentManager.initialFetch();
+            this.currentManager.initialFetch();
         else
             console.error('Failed to fetch selected filter option. "initialFetch" is not a function');
